requirements.txt
# Minimal CrisisFlow dependencies
numpy
pyyaml
